autentificacao/app/requirements.txt
fastapi==0.110.1
sqlmodel==0.0.16
uvicorn[standard]==0.29.0
pymysql==1.1.1
bcrypt==4.1.3
pyjwt==2.8.0